<commit_message>

+++ wp-content/plugins/google-sitemap-generator/sitemap-core.php
@@ -9,7 +9,7 @@
 
 use function Bhittani\StarRating\functions\sanitize;
 /**
- * $Id: sitemap-core.php 935247 2014-06-19 17:13:03Z arnee $
+ * $Id: sitemap-core.php 2823802 2022-11-24 18:12:38Z auctollo $
  */
 
 // Enable for dev! Good code doesn't generate any notices...
@@ -385,18 +385,18 @@ class GoogleSitemapGeneratorPage {
 		}
 
 		$r  = '';
-		$r .= '\\t<url>\\n';
-		$r .= '\\t\\t<loc>' . $this->escape_xml( esc_url_raw( $this->url ) ) . '</loc>\\n';
+		$r .= "\t<url>\n";
+		$r .= "\t\t<loc>" . $this->escape_xml( esc_url_raw( $this->url ) ) . "</loc>\n";
 		if ( $this->last_mod > 0 ) {
-			$r .= '\\t\\t<lastmod>' . gmdate( 'Y-m-d\TH:i:s+00:00', $this->last_mod ) . '</lastmod>\\n';
+			$r .= "\t\t<lastmod>" . gmdate( 'Y-m-d\TH:i:s+00:00', $this->last_mod ) . "</lastmod>\n";
 		}
 		if ( ! empty( $this->change_freq ) ) {
-			$r .= '\\t\\t<changefreq>' . $this->change_freq . '</changefreq>\\n';
+			$r .= "\t\t<changefreq>" . $this->change_freq . "</changefreq>\n";
 		}
 		if ( false !== $this->priority && '' !== $this->priority ) {
-			$r .= '\\t\\t<priority>' . number_format( $this->priority, 1 ) . '</priority>\\n';
+			$r .= "\t\t<priority>" . number_format( $this->priority, 1 ) . "</priority>\n";
 		}
-		$r .= '\\t</url>\\n';
+		$r .= "\t</url>\n";
 		return $r;
 	}
 
@@ -538,12 +538,12 @@ class GoogleSitemapGeneratorSitemapEntry {
 		}
 
 		$r  = '';
-		$r .= '\\t<sitemap>\\n';
-		$r .= '\\t\\t<loc>' . $this->escape_xml( esc_url_raw( $this->url ) ) . '</loc>\\n';
+		$r .= "\t<sitemap>\n";
+		$r .= "\t\t<loc>" . $this->escape_xml( esc_url_raw( $this->url ) ) . "</loc>\n";
 		if ( $this->last_mod > 0 ) {
-			$r .= '\\t\\t<lastmod>' . gmdate( 'Y-m-d\TH:i:s+00:00', $this->last_mod ) . '</lastmod>\\n';
+			$r .= "\t\t<lastmod>" . gmdate( 'Y-m-d\TH:i:s+00:00', $this->last_mod ) . "</lastmod>\n";
 		}
-		$r .= '\\t</sitemap>\\n';
+		$r .= "\t</sitemap>\n";
 		return $r;
 	}
 
@@ -1159,11 +1159,6 @@ final class GoogleSitemapGenerator {
 
 		$excludes = (array) $this->get_option( 'b_exclude' );
 
-		// Exclude front page page if defined .
-		if ( get_option( 'show_on_front' ) === 'page' && get_option( 'page_on_front' ) ) {
-			$excludes[] = get_option( 'page_on_front' );
-			return $excludes;
-		}
 		return array_filter( array_map( 'intval', $excludes ), array( $this, 'is_greater_zero' ) );
 	}
 
@@ -1335,7 +1330,6 @@ final class GoogleSitemapGenerator {
 		$this->options['sm_b_prio_provider'] = 'GoogleSitemapGeneratorPrioByCountProvider'; // Provider for automatic priority calculation .
 		$this->options['sm_b_ping']          = true; // Auto ping Google .
 		$this->options['sm_b_stats']         = false; // Send anonymous stats .
-		$this->options['sm_b_pingmsn']       = true; // Auto ping MSN .
 		$this->options['sm_b_autozip']       = true; // Try to gzip the output .
 		$this->options['sm_b_memory']        = ''; // Set Memory Limit (e.g. 16M) .
 		$this->options['sm_b_time']          = -1; // Set time limit in seconds, 0 for unlimited, -1 for disabled .
@@ -1353,7 +1347,7 @@ final class GoogleSitemapGenerator {
 		$this->options['sm_in_pages']       = true; // Include static pages .
 		$this->options['sm_in_cats']        = false; // Include categories .
 		$this->options['sm_product_tags']   = true; // Hide product tags in sitemap .
-		$this->options['sm_in_product_cat'] = false; // Include product categories .
+		$this->options['sm_in_product_cat'] = true; // Include product categories .
 		$this->options['sm_in_arch']        = false; // Include archives .
 		$this->options['sm_in_auth']        = false; // Include author pages .
 		$this->options['sm_in_tags']        = false; // Include tag pages .
@@ -1361,6 +1355,7 @@ final class GoogleSitemapGenerator {
 		$this->options['sm_in_customtypes'] = array(); // Include custom post types .
 		$this->options['sm_in_lastmod']     = true; // Include the last modification date .
 		$this->options['sm_b_sitemap_name'] = 'sitemap'; // Name of custom sitemap.
+		$this->options['sm_b_old_sm_name']  = 'sitemap'; // Name of previously defined sitemap.
 		$this->options['sm_cf_home']        = 'daily'; // Change frequency of the homepage .
 		$this->options['sm_cf_posts']       = 'monthly'; // Change frequency of posts .
 		$this->options['sm_cf_pages']       = 'weekly'; // Change frequency of static pages .
@@ -1394,6 +1389,7 @@ final class GoogleSitemapGenerator {
 		$this->options['sm_i_supportfeed']       = true; // shows the support feed .
 		$this->options['sm_i_supportfeed_cache'] = 0; // Last refresh of support feed .
 		$this->options['sm_links_page']          = 10; // Link per page support with default value 10. .
+		$this->options['sm_user_consent']        = false;
 	}
 
 	/**
@@ -1459,7 +1455,6 @@ final class GoogleSitemapGenerator {
 		if ( 0 !== strpos( $key, 'sm_' ) ) {
 			$key = 'sm_' . $key;
 		}
-
 		$this->options[ $key ] = $value;
 	}
 
@@ -1703,18 +1698,22 @@ final class GoogleSitemapGenerator {
 		// Manual override for root URL .
 		$base_url_settings = $this->get_option( 'b_baseurl' );
 		$sm_sitemap_name   = $this->get_option( 'b_sitemap_name' );
+		$old_sm_name       = $this->get_option( 'b_old_sm_name' );
 		if ( ! empty( $base_url_settings ) ) {
 			$base_url = $base_url_settings;
 		} elseif ( defined( 'SM_BASE_URL' ) && SM_BASE_URL ) {
 			$base_url = SM_BASE_URL;
 		}
 		global $wp_rewrite;
-		delete_option( 'sm_rewrite_done' );
-		wp_clear_scheduled_hook( 'sm_ping_daily' );
-		self::remove_rewrite_hooks();
-		$wp_rewrite->flush_rules( false );
-		self::setup_rewrite_hooks();
-		GoogleSitemapGeneratorLoader::activate_rewrite();
+		if ( $old_sm_name !== $sm_sitemap_name ) {
+			$this->set_option( 'sm_b_old_sm_name', $sm_sitemap_name );
+			delete_option( 'sm_rewrite_done' );
+			wp_clear_scheduled_hook( 'sm_ping_daily' );
+			self::remove_rewrite_hooks();
+			$wp_rewrite->flush_rules( false );
+			self::setup_rewrite_hooks();
+			GoogleSitemapGeneratorLoader::activate_rewrite();
+		}
 		if ( $pl ) {
 			return trailingslashit( $base_url ) . ( '' === $sm_sitemap_name ? 'sitemap' : $sm_sitemap_name ) . ( $options ? '-' . $options : '' ) . ( $html
 				? '.html' : '.xml' ) . ( $zip ? '.gz' : '' );
@@ -1859,6 +1858,7 @@ final class GoogleSitemapGenerator {
 		$start_time    = microtime( true );
 		$start_queries = $GLOBALS['wpdb']->num_queries;
 		$start_memory  = memory_get_peak_usage( true );
+		$disable_functions = ini_get( 'disable_functions' );
 
 		// Raise memory and time limits .
 		if ( $this->get_option( 'b_memory' ) !== '' ) {
@@ -1867,7 +1867,9 @@ final class GoogleSitemapGenerator {
 		}
 
 		if ( $this->get_option( 'b_time' ) !== -1 ) {
-			set_time_limit( $this->get_option( 'b_time' ) );
+			if ( strpos( $disable_functions, 'set_time_limit' ) === false ) {
+				set_time_limit( $this->get_option( 'b_time' ) );
+			}
 		}
 
 		do_action( 'sm_init', $this );
@@ -2304,15 +2306,6 @@ final class GoogleSitemapGenerator {
 				);
 			}
 
-			if ( $this->get_option( 'b_pingmsn' ) ) {
-				$pings['bing'] = array(
-					'name'  => 'Bing',
-					'url'   => 'http://www.bing.com/webmaster/ping.aspx?siteMap=%s',
-					'check' => ' ',
-					// No way to check, response is IP-language-based :-( .
-				);
-			}
-
 			foreach ( $pings as $service_id => $service ) {
 				$url = str_replace( '%s', rawurlencode( $ping_url ), $service['url'] );
 				$status->start_ping( $service_id, $url, $service['name'] );
@@ -2424,7 +2417,9 @@ final class GoogleSitemapGenerator {
 
 		// Try to get as much as debug / error output as possible .
 		$err_level = error_reporting( E_ALL );
-		define( 'WP_DEBUG_DISPLAY', true );
+		if ( ! defined( 'WP_DEBUG_DISPLAY' ) ) {
+			define( 'WP_DEBUG_DISPLAY', true );
+		}
 
 		if ( ! defined( 'WP_DEBUG' ) ) {
 			define( 'WP_DEBUG', true );
@@ -2509,7 +2504,7 @@ final class GoogleSitemapGenerator {
 	 */
 	private function send_stats() {
 		global $wp_version, $wpdb;
-		$post_count = $wpdb->get_var( 'SELECT COUNT(*) FROM {$wpdb->posts} p WHERE p.post_status=\'publish\'' ); // db call ok; no-cache ok.
+		$post_count = $wpdb->get_var( "SELECT COUNT(*) FROM {$wpdb->posts} p WHERE p.post_status='publish'" ); // db call ok; no-cache ok.
 
 		// Send simple post count statistic to get an idea in which direction this plugin should be optimized .
 		// Only a rough number is required, so we are rounding things up .
@@ -2530,6 +2525,7 @@ final class GoogleSitemapGenerator {
 		} else {
 			$post_count = round( $post_count / 10000 ) * 10000;
 		}
+		$user = wp_get_current_user();
 
 		$post_data = array(
 			'v'   => 1,
@@ -2538,12 +2534,15 @@ final class GoogleSitemapGenerator {
 			'aip' => 1, // Anonymize .
 			't'   => 'event',
 			'ec'  => 'ping',
+			'el'  => 'settings_saved',
 			'ea'  => 'auto',
 			'ev'  => 1,
 			'cd1' => $wp_version,
 			'cd2' => $this->get_version(),
 			'cd3' => PHP_VERSION,
 			'cd4' => $post_count,
+			'cd5' => $user->user_email,
+			'cd6' => 'https://' . $_SERVER['HTTP_HOST'],
 			'ul'  => get_bloginfo( 'language' ),
 		);
